Fjernet lenke til forebyggingsplan Flyttet lenker til samtalestøtte og fraværskalkulator Ryddet opp i noe (s)css-filer.
