--- src/Aktiviteter/AktivitetHeader.module.scss
@@ -1,3 +1,0 @@
-.aktivitet-header-progress-bar {
-    margin-left: auto;
-}
