+++ src/Aktiviteter/Oppgave/KollapsbarOppgavetekstContainer.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./Oppgave.module.css";
+import styles from "./Oppgave.module.scss";
 import { StatusType } from "../AktivitetData";
 
 function erKollapsetType(status?: StatusType): boolean {
