--- src/Aktiviteter/Oppgave/Oppgave.module.css
@@ -1,83 +1,0 @@
-.oppgaveblokk {
-    --oppgaveblokk-farge: var(--a-surface-selected);
-    --oppgaveblokk-padding: 1rem;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    padding: var(--oppgaveblokk-padding);
-    margin-bottom: 1rem;
-    background-color: var(--oppgaveblokk-farge);
-    border-radius: var(--a-border-radius-medium);
-}
-
-/* Gjør at oppgave får samme padding over og under
- * ved å fjerne margin på siste element i oppgaveblokka */
-.oppgaveinnhold > :last-child {
-    margin-bottom: 0;
-}
-
-.oppgaveinnhold .tittelContainer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    flex-wrap: wrap;
-}
-
-.oppgaveinnhold .tittelContainer .status {
-    margin-left: auto;
-    white-space: nowrap;
-}
-
-.oppgaveinnhold .kollapsetOppgavetekst {
-    min-height: 5rem;
-    max-height: 7rem;
-    overflow: hidden;
-    position: relative;
-}
-
-.oppgaveinnhold .oppgavetekstOverlayGradient {
-    display: flex;
-    justify-content: center;
-    align-items: end;
-}
-
-.oppgaveinnhold .synligOppgavetekst .oppgavetekstOverlayGradient {
-    border-top: thin solid var(--a-gray-400);
-    padding-top: calc(var(--oppgaveblokk-padding) / 2);
-}
-.oppgaveinnhold .kollapsetOppgavetekst .oppgavetekstOverlayGradient {
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    background: linear-gradient(
-        180deg,
-        transparent -15%,
-        var(--oppgaveblokk-farge) 45%
-    );
-    pointer-events: none;
-}
-
-.statusknapp {
-    pointer-events: all;
-    background-color: var(--oppgaveblokk-farge);
-    --a-surface-neutral-subtle-hover: var(--a-blue-100);
-}
-
-.helknapp {
-    width: 15.5rem;
-}
-
-.halvknapp {
-    width: 7.5rem;
-    margin: 0.5rem;
-    margin-bottom: 0;
-}
-
-.multiknappholder {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
-}
