+++ src/Aktiviteter/Oppgave/Oppgave.module.scss
@@ -1,0 +1,79 @@
+/* Gjør at oppgave får samme padding over og under
+ * ved å fjerne margin på siste element i oppgaveblokka */
+.oppgaveblokk {
+  --oppgaveblokk-farge: var(--a-surface-selected);
+  --oppgaveblokk-padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  padding: var(--oppgaveblokk-padding);
+  margin-bottom: 1rem;
+  background-color: var(--oppgaveblokk-farge);
+  border-radius: var(--a-border-radius-medium);
+
+  .oppgaveinnhold {
+    > :last-child {
+      margin-bottom: 0;
+    }
+    .tittelContainer {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      flex-wrap: wrap;
+      .status {
+        margin-left: auto;
+        white-space: nowrap;
+      }
+    }
+    .kollapsetOppgavetekst {
+      min-height: 5rem;
+      max-height: 7rem;
+      overflow: hidden;
+      position: relative;
+      .oppgavetekstOverlayGradient {
+        position: absolute;
+        top: 0;
+        left: 0;
+        bottom: 0;
+        right: 0;
+        background: linear-gradient(
+          180deg,
+          transparent -15%,
+          var(--oppgaveblokk-farge) 45%
+        );
+        pointer-events: none;
+      }
+    }
+    .oppgavetekstOverlayGradient {
+      display: flex;
+      justify-content: center;
+      align-items: end;
+    }
+    .synligOppgavetekst {
+      .oppgavetekstOverlayGradient {
+        border-top: thin solid var(--a-gray-400);
+        padding-top: calc(var(--oppgaveblokk-padding) / 2);
+      }
+    }
+
+    .statusknapp {
+      pointer-events: all;
+      background-color: var(--oppgaveblokk-farge);
+      --a-surface-neutral-subtle-hover: var(--a-blue-100);
+    }
+    .helknapp {
+      width: 15.5rem;
+    }
+    .halvknapp {
+      width: 7.5rem;
+      margin: 0.5rem;
+      margin-bottom: 0;
+    }
+    .multiknappholder {
+      display: flex;
+      flex-wrap: wrap;
+      align-items: center;
+      justify-content: center;
+    }
+  }
+}
