--- src/Forside/AndreForebyggendeVerktøy/andreForebyggendeVerktøy.module.scss
@@ -1,17 +1,0 @@
-@use "src/styles/screensize";
-
-.andreForebyggendeVerktøy {
-  display: flex;
-  flex-direction: column;
-  gap: var(--a-spacing-4);
-
-  .panelGrid {
-    display: flex;
-    flex-direction: column;
-    gap: var(--a-spacing-4);
-
-    @media (min-width: screensize.$screen-lg) {
-      flex-direction: row;
-    }
-  }
-}
