+++ src/Forside/Forside.tsx
@@ -7,8 +7,6 @@ import { useOrgnr } from "../hooks/useOrgnr";
 import { Alert } from "@navikt/ds-react";
 import { tomtDataobjekt } from "../integrasjoner/aggregert-statistikk-api";
 import { leggTilBedriftPåUrl } from "../utils/navigasjon";
-import { NyttVerktoyTilDeg } from "./NyttVerktøyTilDeg/NyttVerktoyTilDeg";
-import { AndreForebyggendeVerktoy } from "./AndreForebyggendeVerktøy/AndreForebyggendeVerktoy";
 import { RelaterteTjenester } from "./RelaterteTjenester/RelaterteTjenester";
 import { Sykefraværsstatistikk } from "./Sykefraværsstatistikk/Sykefraværsstatistikk";
 import { KontaktOss } from "./KontaktOss/KontaktOss";
@@ -19,7 +17,6 @@ import Aktiviteter from "../Aktiviteter/Aktiviteter";
 
 export interface ForsideProps {
   samtalestøtteUrl: string;
-  forebyggingsplanUrl: string;
   sykefraværsstatistikkUrl: string;
   kontaktOssUrl: string;
   kjørerMockApp: boolean;
@@ -65,13 +62,11 @@ export const Forside = (props: ForsideProps) => {
       <div className={styles.forside}>
         {props.children}
         {sykefraværsstatistikkEllerBannerHvisError}
-        <Aktiviteter />
-        <NyttVerktoyTilDeg href={props.forebyggingsplanUrl} />
+        <Aktiviteter samtalestøtteUrlMedOrgnr={samtalestøtteUrlMedOrgnr} />
         {fiaSamarbeidsstatus.status === RestStatus.Suksess &&
           fiaSamarbeidsstatus.data.samarbeid === "I_SAMARBEID" && (
             <FiaSamarbeidsstatus status={fiaSamarbeidsstatus.data.samarbeid} />
           )}
-        <AndreForebyggendeVerktoy href={samtalestøtteUrlMedOrgnr} />
         <RelaterteTjenester />
         <KontaktOss kontaktOssUrl={props.kontaktOssUrl} />
       </div>
