--- src/Forside/NyttVerktøyTilDeg/nyttVerktøyTilDeg.module.scss
@@ -1,22 +1,0 @@
-@use "src/styles/screensize";
-
-.nyttVerktøyTilDeg {
-  background-color: var(--a-deepblue-50);
-  padding: var(--a-spacing-5);
-  border-radius: 4px;
-
-  .innhold {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    .enhet {
-      flex-basis: 50%
-    }
-
-    @media (min-width: screensize.$screen-md) {
-      gap: var(--a-spacing-4);
-      flex-direction: row;
-    }
-  }
-}
